refactor: refactoring the guest-request component
