+++ frontend/src/app/giftList/[id]/(components)/giftsListDisplay/guest-list.tsx
