+++ frontend/src/app/giftList/[id]/(components)/giftsListDisplay/owner-list.tsx
